refractor App.jsx in order to clean the code, created the hook, useCartManager in oder to make the code more modular
